File renames and moves.  And organizational commit.
