+++ DataRepo/tests/management/test_load_protocols.py
@@ -16,8 +16,6 @@ from DataRepo.utils.exceptions import (
     RequiredColumnValues,
 )
 
-# TODO: Rename to test_load_protocols.py
-
 
 @tag("protocols")
 class ProtocolLoadingTests(TracebaseTestCase):
