+++ DataRepo/tests/management/test_load_tissues.py
@@ -9,8 +9,6 @@ from DataRepo.utils.exceptions import (
     RequiredColumnValues,
 )
 
-# TODO: Rename to test_load_tissues.py
-
 
 @tag("tissues")
 class TissueLoadingTests(TracebaseTestCase):
